refactor analysis - improve tile processing

+++ server/src/main/scala/kpn/server/analyzer/engine/tiles/domain/TileId.scala
@@ -1,6 +1,6 @@
 package kpn.server.analyzer.engine.tiles.domain
 
-case class TileId(z: Long, x: Long, y: Long) {
+case class TileId(z: Int, x: Int, y: Int) {
   def name: String = {
     s"$z-$x-$y"
   }
